Set max length of phone number in enter phone number screen

+++ app/src/main/java/com/anvipo/angram/presentationLayer/userStories/authUserStory/screens/enterAuthCode/presenter/EnterAuthCodePresenterImp.kt
@@ -72,7 +72,7 @@ class EnterAuthCodePresenterImp(
 
     override fun onGetExpectedCodeLength(expectedCodeLength: UInt) {
         this.expectedCodeLength = expectedCodeLength
-        viewState.setMaxLengthOfEditText(this.expectedCodeLength)
+        viewState.setMaxLengthOfEditText(this.expectedCodeLength.toInt())
     }
 
     override fun onGetEnteredPhoneNumber(enteredPhoneNumber: String) {
